+++ scripts/03_02_get_daily_dist_to_fire.R
@@ -1,4 +1,6 @@
-if (!dir.exists(file.path(path_fire, "distance_to_fire", "grid_dist_fire"))) dir.create(file.path(path_fire, "distance_to_fire", "grid_dist_fire"), recursive = T)
+if (!dir.exists(file.path(path_fire, "distance_to_fire", "grid_dist_fire"))) {
+  dir.create(file.path(path_fire, "distance_to_fire", "grid_dist_fire"), recursive = T)
+}
 
 #-------------------------------------------------------------------------------
 # Get Distance to Fire Over 10km Grid
@@ -8,51 +10,55 @@ if (!dir.exists(file.path(path_fire, "distance_to_fire", "grid_dist_fire"))) dir
 #-------------------------------------------------------------------------------
 years = 2006:2020
 
-# Load 10km grid
-grid = readRDS(file.path(path_boundaries, "grid.RDS"))
-poly_grid = gBuffer(grid, byid=T, width=5000, capStyle="SQUARE") #slow
-
-# loop through years
-fire_files = list.files(path_fire)
-for (i in 1:length(years)) {
+if (!all(file.exists(file.path(path_fire, "distance_to_fire", "grid_dist_fire", 
+                               paste0("dist_to_fire_", 2006:2020, ".RDS"))))) {
   
-  # get year
-  year = years[i]
+  # Load 10km grid
+  grid = readRDS(file.path(path_boundaries, "grid.RDS"))
+  poly_grid = gBuffer(grid, byid=T, width=5000, capStyle="SQUARE") #slow
   
-  # get list of days in that year
-  days = seq(as.Date(paste0(year, "-01-01")), 
-             as.Date(paste0(year, "-12-31")), "days")
-  days = gsub("-", "", as.character(days))
-  out = as.list(rep(NA, length(days)))
-  
-  # loop through days in the year
-  for (j in 1:length(days)) {
+  # loop through years
+  fire_files = list.files(path_fire)
+  for (i in 1:length(years)) {
     
-    #figure out file to open
-    file = paste0("hms_fire", days[j], ".shp")
+    # get year
+    year = years[i]
     
-    if (file %in% fire_files) {
-      fires = readOGR(file.path(path_fire, paste0("points/hms_fire", days[j], ".shp")), 
-                      paste0("hms_fire", days[j]), verbose=F)
-      crs(fires) = "+proj=longlat +datum=WGS84 +no_defs"
-      fires = fires[fires$Lon > -144 & fires$Lon < -32 & 
-                      fires$Lat > -7 & fires$Lat < 70, ] #limit to north america
-      fires = spTransform(fires, crs(grid))
+    # get list of days in that year
+    days = seq(as.Date(paste0(year, "-01-01")), 
+               as.Date(paste0(year, "-12-31")), "days")
+    days = gsub("-", "", as.character(days))
+    out = as.list(rep(NA, length(days)))
+    
+    # loop through days in the year
+    for (j in 1:length(days)) {
       
-      dist = get.knnx(coordinates(fires), coordinates(grid), k=1)
+      #figure out file to open
+      file = paste0("hms_fire", days[j], ".shp")
       
-      df = data.frame(id=grid$ID, date=days[j], km_dist=dist$nn.dist/1000)
-    } else {
-      df = data.frame(id=grid$ID, date=days[j], km_dist=NA)
+      if (file %in% fire_files) {
+        fires = readOGR(file.path(path_fire, paste0("points/hms_fire", days[j], ".shp")), 
+                        paste0("hms_fire", days[j]), verbose=F)
+        crs(fires) = "+proj=longlat +datum=WGS84 +no_defs"
+        fires = fires[fires$Lon > -144 & fires$Lon < -32 & 
+                        fires$Lat > -7 & fires$Lat < 70, ] #limit to north america
+        fires = spTransform(fires, crs(grid))
+        
+        dist = get.knnx(coordinates(fires), coordinates(grid), k=1)
+        
+        df = data.frame(id=grid$ID, date=days[j], km_dist=dist$nn.dist/1000)
+      } else {
+        df = data.frame(id=grid$ID, date=days[j], km_dist=NA)
+      }
+      
+      out[[j]] = df
+      if (j%%50==0) {print(j)}
     }
     
-    out[[j]] = df
-    if (j%%50==0) {print(j)}
+    out = rbindlist(out)
+    saveRDS(out, file.path(path_fire, "distance_to_fire", "grid_dist_fire", paste0("dist_to_fire_", year, ".RDS")))
+    
+    print(year)
+    
   }
-  
-  out = rbindlist(out)
-  saveRDS(out, file.path(path_fire, "distance_to_fire", "grid_dist_fire", paste0("dist_to_fire_", year, ".RDS")))
-  
-  print(year)
-  
 }
